requirements/requirements.txt
pandas
numpy
regex
spacy
selenium
dateparser
psycopg2-binary
SQLAlchemy
beautifulsoup4
fake-useragent
requests
webdriver-manager
docker